+++ src/node/rpc/network_identity_subsystem.h
@@ -144,7 +144,7 @@ namespace ccf
     {
       if (fetch_status.load() != FetchStatus::Done)
       {
-        throw std::logic_error(fmt::format(
+        throw EndorsementsNotFetchedError(fmt::format(
           "COSE endorsements chain requested for seqno {} but endorsement "
           "fetching has not been completed yet",
           seqno));
@@ -186,7 +186,7 @@ namespace ccf
     {
       if (fetch_status.load() != FetchStatus::Done)
       {
-        throw std::logic_error(fmt::format(
+        throw EndorsementsNotFetchedError(fmt::format(
           "Trusted key requested for seqno {} but the fetching has "
           "not been completed yet",
           seqno));
@@ -217,7 +217,7 @@ namespace ccf
     {
       if (fetch_status.load() != FetchStatus::Done)
       {
-        throw std::logic_error(
+        throw EndorsementsNotFetchedError(
           "Trusted keys requested but endorsements/key fetching has not "
           "completed yet");
       }